+++ src/main/java/com/hhp/ecommerce/presentation/dto/ProductResponse.java
@@ -1,4 +1,4 @@
-package com.hhp.ecommerce.interfaces.dto;
+package com.hhp.ecommerce.presentation.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProductSalesResponse {
+public class ProductResponse {
 	private String productId;
 	private String name;
 	private int price;
-	private int totalSales;
+	private int stock;
 }
